refactor: 메인의 각 영역을 컴포넌트로 분리
